+++ publish/CHANGELOG.md
@@ -1,3 +1,7 @@
+- v1.13
+  - Adds support for value "snap" to y-coordinate for `spawn` field.
+  - Fixed for the new game version.
+
 - v1.12
   - Adds support for parameter "<key_>" to get value of a global key.
   - Adds wild card matching to string filters.